% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{httpcode-package}
\alias{httpcode}
\alias{httpcode-package}
\title{Explain the meaning of HTTP status codes}
\description{
Explain the meaning of HTTP status codes
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

