% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cat_for_status}
\alias{cat_for_status}
\alias{dog_for_status}
\title{Use cat pictures for various HTTP status codes}
\usage{
cat_for_status(code)

dog_for_status(code)
}
\arguments{
\item{code}{An http status code}
}
\description{
Use cat pictures for various HTTP status codes
}
\examples{
\donttest{
cat_for_status(100)
dog_for_status(100)
cat_for_status(301)
dog_for_status(301)
cat_for_status(400)
dog_for_status(400)
}
}

