% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cat_for_status}
\alias{cat_for_status}
\alias{dog_for_status}
\title{Use cat or dog pictures for various HTTP status codes}
\usage{
cat_for_status(code, browse = FALSE)

dog_for_status(code, browse = FALSE)
}
\arguments{
\item{code}{An http status code}

\item{browse}{(logical) If TRUE, opens image in default browser. If FALSE,
returns the URL of the image.}
}
\value{
Opens image in your default browser, or returns URL
}
\description{
Use cat or dog pictures for various HTTP status codes
}
\examples{
\donttest{
# give back url
cat_for_status(100)
dog_for_status(100)
cat_for_status(301)
dog_for_status(301)
cat_for_status(400)
dog_for_status(400)

# open image in default browser
cat_for_status(400, browse=TRUE)

# not found
cat_for_status(555)
}
}

