% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpcode-package.R
\docType{package}
\name{httpcode-package}
\alias{httpcode}
\alias{httpcode-package}
\title{Explain the meaning of HTTP status codes}
\description{
Explain the meaning of HTTP status codes
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

